use sanic to run background tasks

+++ utils/FileUtils.py
@@ -86,6 +86,7 @@ def DeleteFiles(files: list[str], types: list[str]) -> None:
     """
     for file in files:
         for file_type in types:
-            delete_file = f'{file}.{file_type}'
-            logging.debug('deleting file: %s' % delete_file)
-            os.remove(delete_file)
+            if os.path.exists(file + '.' + file_type):
+                delete_file = f'{file}.{file_type}'
+                logging.debug('deleting file: %s' % delete_file)
+                os.remove(delete_file)
